Fix #6 Prevent saving empty currency and empty name of asset
