+++ src/pages/new/ui/SearchSelect/index.ts
@@ -1,0 +1,1 @@
+export { SearchSelect } from "./SearchSelect";
