+++ src/shared/currencies/currenciesList.ts
@@ -1,0 +1,7 @@
+import { currenciesByCode } from "./currenciesByCode";
+
+export const currenciesList = Object.entries(currenciesByCode).map(
+  ([code, nameEn]) => {
+    return { code: code, en: nameEn };
+  }
+);
