+++ src/shared/currencies/index.ts
@@ -1,0 +1,1 @@
+export { currenciesByCode } from "./currenciesByCode";
